Move overridables into a separate map, so that the template can override them.
